lab1. Solving system of equation using the Gaussian elimination. This program solves dependent/independent and consistent systems NxM, where N can be not equal to M.
lab2. Building the convex hull with the Chan's algorythm (combination of the Graham scan algorythm and the Gift wrapping algorythm).

lab3. Parallel Gaussian elimination. Works only for consistent independent system.

lab4. "The hedgehog picks up apples" with OpenGL 4.0.

lab5. Realization of the two-dimensional array on the base of class Tree.